iceberg

+++ wealth-data-node/wealth-data-node-iceberg/src/main/java/com/enhui/connect/test/IcebergWriter.java
@@ -9,9 +9,14 @@ import java.io.IOException;
 import java.util.Arrays;
 import java.util.UUID;
 import org.apache.hadoop.conf.Configuration;
+import org.apache.iceberg.AppendFiles;
+import org.apache.iceberg.DataFile;
+import org.apache.iceberg.DeleteFile;
 import org.apache.iceberg.FileFormat;
+import org.apache.iceberg.RowDelta;
 import org.apache.iceberg.Schema;
 import org.apache.iceberg.Table;
+import org.apache.iceberg.Transaction;
 import org.apache.iceberg.catalog.TableIdentifier;
 import org.apache.iceberg.data.GenericAppenderFactory;
 import org.apache.iceberg.data.GenericRecord;
@@ -26,8 +31,8 @@ import org.apache.iceberg.types.Types;
 
 public class IcebergWriter {
 
-  static boolean isPartition = false;
-  static boolean isDelta = false;
+  static boolean isPartition = true;
+  static boolean isDelta = true;
 
   public static void main(String[] args) throws IOException {
     IcebergWriter icebergWriter = new IcebergWriter();
@@ -41,44 +46,45 @@ public class IcebergWriter {
     String namespace = "icebergdb";
     String tableName = "writer_test_" + isPartition + "_" + isDelta;
     TableIdentifier name = TableIdentifier.of(namespace, tableName);
+    boolean isUpdate = true;
+    if (!isUpdate) {
+      catalog.dropTable(name);
+    }
 
-    catalog.dropTable(name);
-
-    Table table = IcebergClient.createOrLoadTable(catalog, name, isPartition);
-
-    // 事务，先插入，在删除(删除不能针对行)
-    //    Transaction transaction = table.newTransaction();
-    //    DataFile dataFile = IcebergClient.getDataFileWithRecords(table, table.schema(), 30);
-    //
-    //    transaction.newAppend().appendFile(dataFile).commit();
-    //    transaction.newDelete().deleteFromRowFilter(Expressions.in("id",20,30)).commit();
-    //    transaction.commitTransaction();
+    Table table = IcebergClient.createOrLoadTable(catalog, name, isPartition, isDelta);
 
     TaskWriter<Record> taskWriter =
         icebergWriter.getUnpartitionedWriter(table, FileFormat.PARQUET, isPartition, isDelta);
-    for (int i = 0; i < 5000; i++) {
-      taskWriter.write(icebergWriter.getRecord(table, i));
-    }
-    taskWriter.write(icebergWriter.getRecord(table, 2));
-    taskWriter.write(icebergWriter.getRecord(table, 1));
+    taskWriter.write(icebergWriter.getRecord(table, 1, isUpdate));
+    taskWriter.write(icebergWriter.getRecord(table, 2, isUpdate));
+    taskWriter.write(icebergWriter.getRecord(table, 1, isUpdate));
     taskWriter.close();
     final WriteResult complete = taskWriter.complete();
-    System.out.println(Arrays.toString(complete.referencedDataFiles()));
-    System.out.println(Arrays.toString(complete.deleteFiles()));
-    System.out.println(Arrays.toString(complete.dataFiles()));
+    System.out.println("\nreferencedDataFiles：" + Arrays.toString(complete.referencedDataFiles()));
+    System.out.println("\ndeleteFiles：" + Arrays.toString(complete.deleteFiles()));
+    System.out.println("\ndataFiles：" + Arrays.toString(complete.dataFiles()));
 
-    table.refresh();
+    final Transaction transaction = table.newTransaction();
+    for (DataFile dataFile : complete.dataFiles()) {
+      AppendFiles appendFiles = transaction.newAppend();
+      appendFiles.appendFile(dataFile).commit();
+    }
+    for (DeleteFile deleteFile : complete.deleteFiles()) {
+      RowDelta rowDelta = transaction.newRowDelta();
+      rowDelta.addDeletes(deleteFile).commit();
+    }
+    transaction.commitTransaction();
 
     IcebergClient.selectAndPrint(table);
   }
 
-  public Record getRecord(Table table, int i) {
+  public Record getRecord(Table table, int i, boolean idUpdate) {
     GenericRecord record = GenericRecord.create(table.schema());
     for (Types.NestedField column : table.schema().columns()) {
       if (column.type() == Types.IntegerType.get()) {
         record.setField(column.name(), i);
       } else if (column.type() == Types.StringType.get()) {
-        record.setField(column.name(), String.valueOf(i));
+        record.setField(column.name(), String.valueOf(idUpdate ? i + 1 : i));
       } else {
         throw new RuntimeException("没处理的类型，换成int 和 string 测试");
       }
